+++ lithops/serverless/backends/aws_lambda/entry_point.py
@@ -20,7 +20,7 @@ from lithops.version import __version__
 from lithops.utils import setup_lithops_logger
 from lithops.worker import function_handler
 from lithops.worker import function_invoker
-from lithops.worker.utils import get_runtime_preinstalls
+from lithops.worker.utils import get_runtime_metadata
 
 logger = logging.getLogger('lithops.worker')
 
@@ -31,14 +31,14 @@ def lambda_handler(event, context):
 
     setup_lithops_logger(event.get('log_level', logging.INFO))
 
-    if 'get_preinstalls' in event:
-        logger.info("Lithops v{} - Generating metadata".format(__version__))
-        return get_runtime_preinstalls()
+    if 'get_metadata' in event:
+        logger.info(f"Lithops v{__version__} - Generating metadata")
+        return get_runtime_metadata()
     elif 'remote_invoker' in event:
-        logger.info("Lithops v{} - Starting AWS Lambda invoker".format(__version__))
+        logger.info(f"Lithops v{__version__} - Starting AWS Lambda invoker")
         function_invoker(event)
     else:
-        logger.info("Lithops v{} - Starting AWS Lambda execution".format(__version__))
+        logger.info(f"Lithops v{__version__} - Starting AWS Lambda execution")
         function_handler(event)
 
     return {"Execution": "Finished"}